Added alarm probing. Fixed position overflow.

+++ alarm_blinds_v1/BlindsServoSM.h
@@ -27,8 +27,8 @@ class CBlindsServoSM {
     int m_iServoPin;
 
     // Position and direction info. Position measured in us travelled in that direction.
-    int m_iCurrentPosition; // When iPosition is 0, blinds are down.
-    int m_iUpPosition; // When iPosition = iUpPosition, blinds are up.
+    long m_iCurrentPosition; // When iPosition is 0, blinds are down.
+    long m_iUpPosition; // When iPosition = iUpPosition, blinds are up.
     int m_iServoDir; // Current direction servo is moving in 
 
     int m_iDownCmd;
